+++ server/ts/mission_upload.ts
@@ -5,11 +5,12 @@ import * as crypto from 'crypto';
 import { MisFile, MisParser, MissionElementScriptObject, MissionElementSimGroup, MissionElementType } from './io/mis_parser';
 import { Config } from './config';
 import hxDif from '../lib/hxDif';
-import { IGNORE_MATERIALS, IMAGE_EXTENSIONS, Mission } from './mission';
+import { IGNORE_MATERIALS, IMAGE_EXTENSIONS, Mission, MissionDoc } from './mission';
 import { Util } from './util';
 import { DtsParser } from './io/dts_parser';
 import { db, keyValue } from './globals';
-import sharp from 'sharp';
+import { createPack, createPackThumbnail, PackDoc } from './pack';
+import { AccountDoc } from './account';
 
 /** Stores a list of currently ongoing uploads that are waiting to be submitted. */
 export const ongoingUploads = new Map<string, MissionUpload>();
@@ -22,7 +23,20 @@ setInterval(async () => {
 	}
 }, 1000 * 60 * 60 * 24);
 
-/** Represents a mission upload process. Handles processing, verification and submission of the uploaded mission. */
+/** Stores data for each mission uploaded through an archive. */
+interface MissionGroup {
+	misFile: jszip.JSZipObject,
+	misFilePath: string,
+	missionInfo: MissionElementScriptObject,
+	thumbnailFile: jszip.JSZipObject,
+	interiorDependencies: Set<string>,
+	skyDependencies: Set<string>,
+	shapeDependencies: Set<string>,
+	/** Stores the final, normalized directory structure for this level. */
+	normalizedDirectory: Map<string, jszip.JSZipObject>
+}
+
+/** Represents a mission upload process. Handles processing, verification and submission of the uploaded missions. */
 export class MissionUpload {
 	/** A ID to uniquely identify this upload. */
 	id = Math.random().toString(); // Plenty safe for this usecase 🥴
@@ -33,21 +47,13 @@ export class MissionUpload {
 	problems = new Set<string>();
 	/** Stores a list of warnings regarding mission upload, which are things to pay attention to but not things that will prevent submission. */
 	warnings = new Set<string>();
-	misFile: MisFile;
-	/** The path to the .mis file within the .zip directory. */
-	misFilePath: string;
-	missionInfo: MissionElementScriptObject;
-	interiorDependencies = new Set<string>();
-	skyDependencies = new Set<string>();
-	shapeDependencies = new Set<string>();
-	/** Stores the final, normalized directory structure for this level. */
-	normalizedDirectory = new Map<string, jszip.JSZipObject>();
+	groups: MissionGroup[] = [];
 
 	constructor(zip: jszip) {
 		this.zip = zip;
 	}
 
-	/** Processes the uploaded level, building the dependency tree while finding any problems present.
+	/** Processes the uploaded missions, building the dependency tree while finding any problems present.
 	 * 
 	 * General explanation of the archive resolution process:
 	 * Every time we try to find files in the uploaded archive, we view the archive as "flattened", meaning we assume that there are no subdirectories and
@@ -60,16 +66,20 @@ export class MissionUpload {
 	*/
 	async process() {
 		this.sanitizeArchive();
+		this.createMissionGroups();
 
-		let continueProcess = await this.checkMis();
-		if (!continueProcess) return; // If the mis wasn't valid, no need to even continue
-
-		await this.checkInteriors();
-		await this.checkSky();
-		await this.checkShapes();
+		// Now, process every mission separately
+		for (let group of this.groups) {
+			let continueProcess = await this.checkMis(group);
+			if (!continueProcess) continue; // If the mis wasn't valid, no need to even continue
+	
+			await this.checkInteriors(group);
+			await this.checkSky(group);
+			await this.checkShapes(group);
+		}
 
 		// Determine all files that are part of the archive but won't be included in the final directory
-		let includedFiles = new Set(this.normalizedDirectory.values());
+		let includedFiles = new Set(this.groups.map(x => [...x.normalizedDirectory.values()]).flat());
 		let excludedFiles: string[] = [];
 
 		for (let filePath in this.zip.files) {
@@ -94,7 +104,7 @@ export class MissionUpload {
 		}
 	}
 
-	async checkMis() {
+	createMissionGroups() {
 		let misFiles: jszip.JSZipObject[] = [];
 
 		// Find all .mis files in the archive
@@ -102,28 +112,41 @@ export class MissionUpload {
 			if (filePath.toLowerCase().endsWith('.mis')) misFiles.push(this.zip.files[filePath]);
 		}
 
-		// We expect a single level to be uploaded, so abort if we find more than one .mis file.
-		if (misFiles.length !== 1) {
-			this.problems.add(`The archive must contain exactly one .mis file (found ${misFiles.length}).`);
-			return false;
+		if (misFiles.length === 0) {
+			this.problems.add(`The archive must contain at least one .mis file.`);
 		}
 
-		this.misFilePath = misFiles[0].name;
-		this.normalizedDirectory.set(path.posix.join('missions', Util.getFileName(this.misFilePath)), misFiles[0]); // Set the .mis file to be in the missions/ directory
+		for (let file of misFiles) {
+			let group: MissionGroup = {
+				misFile: file,
+				misFilePath: file.name,
+				missionInfo: null,
+				thumbnailFile: null,
+				interiorDependencies: new Set(),
+				skyDependencies: new Set(),
+				shapeDependencies: new Set(),
+				normalizedDirectory: new Map()
+			};
+			this.groups.push(group);
+		}
+	}
 
-		let text = await misFiles[0].async('text'); // This might fuck up in extremely rare cases due to encoding stuff (because missions aren't UTF-8)
+	async checkMis(group: MissionGroup) {
+		group.normalizedDirectory.set(path.posix.join('missions', Util.getFileName(group.misFilePath)), group.misFile); // Set the .mis file to be in the missions/ directory
+
+		let text = await group.misFile.async('text'); // This might fuck up in extremely rare cases due to encoding stuff (because missions aren't UTF-8)
 		let misFile: MisFile;
 
 		// Check if this .mis file is already present in the database and if so, abort
 		let misHash = crypto.createHash('sha256').update(text).digest('base64');
 		let existing = await db.missions.findOne({ misHash: misHash });
 		if (existing) {
-			this.problems.add("Duplicate: This mission has already been uploaded.");
+			this.problems.add(`Duplicate: Mission ${group.misFilePath} has already been uploaded.`);
 			return false;
 		}
 
 		if (!text.includes('//--- OBJECT WRITE BEGIN ---') || !text.includes('//--- OBJECT WRITE END ---')) {
-			this.problems.add(`Your .mis file does not contain both "//--- OBJECT WRITE BEGIN ---" and "//--- OBJECT WRITE END ---", which are necessary for the level to behave correctly in the editor. Please add them around the outer MissionGroup element.`);
+			this.problems.add(`${group.misFilePath} does not contain both "//--- OBJECT WRITE BEGIN ---" and "//--- OBJECT WRITE END ---", which are necessary for the level to behave correctly in the editor. Please add them around the outer MissionGroup element.`);
 			return false;
 		}
 
@@ -131,24 +154,23 @@ export class MissionUpload {
 			let parser = new MisParser(text);
 			misFile = parser.parse();
 		} catch (e) {
-			this.problems.add(`Could not parse ${this.misFilePath}.`);
+			this.problems.add(`Could not parse ${group.misFilePath}.`);
 			return false;
 		}
 
-		this.misFile = misFile;
-		await this.traverseMis(misFile.root); // Start scanning all elements in the .mis file to find possible dependencies
+		await this.traverseMis(group, misFile.root); // Start scanning all elements in the .mis file to find possible dependencies
 
-		if (!this.missionInfo) {
-			this.problems.add("The mission does not contain a MissionInfo ScriptObject.");
+		if (!group.missionInfo) {
+			this.problems.add(`${group.misFilePath} does not contain a MissionInfo ScriptObject.`);
 		} else {
 			let missing: string[] = [];
-			if (!this.missionInfo.name) missing.push("name");
-			if (!this.missionInfo.artist) missing.push("artist");
+			if (!group.missionInfo.name) missing.push("name");
+			if (!group.missionInfo.artist) missing.push("artist");
 
-			if (missing.length > 0) this.problems.add(`The mission info is missing the following fields: ${missing.join(', ')}.`);
+			if (missing.length > 0) this.problems.add(`The mission info in ${group.misFilePath} is missing the following fields: ${missing.join(', ')}.`);
 		}
 
-		let startsWith = Util.removeExtension(misFiles[0].name);
+		let startsWith = Util.removeExtension(group.misFile.name);
 		// Determine all potential image thumbnail names
 		let potentialNames = IMAGE_EXTENSIONS.map(x => (startsWith + x).toLowerCase());
 		let imageFound = false;
@@ -158,47 +180,49 @@ export class MissionUpload {
 				imageFound = true;
 
 				let buffer = await this.zip.files[filePath].async('nodebuffer');
-				let metadata = await sharp(buffer).metadata();
+				let dimensions = await Util.getImageDimensions(buffer);
 
-				if (Math.max(metadata.width, metadata.height) < 128) {
-					this.warnings.add(`The image thumbnail contained with the level has a very low resolution (${metadata.width}x${metadata.height}). Please consider upping its resolution or including a .prev.png with your level.`);
+				if (Math.max(dimensions.width, dimensions.height) < 128) {
+					this.warnings.add(`The image thumbnail contained with the level ${group.misFilePath} has a very low resolution (${dimensions.width}x${dimensions.height}). Please consider upping its resolution or including a .prev.png with your level.`);
 				}
+
+				group.thumbnailFile = this.zip.files[filePath];
 			}
 
 			// Include all files as a dependency who start with the .mis file name and end with an image extension. So, level.mis would cause the inclusion of level.png, level.jpg and also level.prev.png and things like that.
 			let fileName = Util.getFileName(filePath);
-			if (fileName.toLowerCase().startsWith(Util.removeExtension(Util.getFileName(misFiles[0].name).toLowerCase())) && IMAGE_EXTENSIONS.includes(path.extname(fileName.toLowerCase()))) {
-				this.normalizedDirectory.set(path.posix.join('missions', fileName), this.zip.files[filePath]);
+			if (fileName.toLowerCase().startsWith(Util.removeExtension(Util.getFileName(group.misFile.name).toLowerCase())) && IMAGE_EXTENSIONS.includes(path.extname(fileName.toLowerCase()))) {
+				group.normalizedDirectory.set(path.posix.join('missions', fileName), this.zip.files[filePath]);
 			}
 		}
 		if (!imageFound) {
-			this.problems.add("The level is missing an image thumbnail.");
+			this.problems.add(`The level ${group.misFilePath} is missing an image thumbnail.`);
 		}
 
 		return true;
 	}
 
-	async traverseMis(simGroup: MissionElementSimGroup) {
+	async traverseMis(missionGroup: MissionGroup, simGroup: MissionElementSimGroup) {
 		for (let element of simGroup.elements) {
-			if (element._type === MissionElementType.ScriptObject && element._name === 'MissionInfo' && !this.missionInfo) {
-				this.missionInfo = element;
+			if (element._type === MissionElementType.ScriptObject && element._name === 'MissionInfo' && !missionGroup.missionInfo) {
+				missionGroup.missionInfo = element;
 
 				// Check for a music dependency
-				if (element.music) await this.registerDependency(path.posix.join('sound/music', element.music), 'exact', this.misFilePath);
+				if (element.music) await this.registerDependency(missionGroup, path.posix.join('sound/music', element.music), 'exact', missionGroup.misFilePath);
 			} else if (element._type === MissionElementType.SimGroup) {
-				await this.traverseMis(element); // Recurse
+				await this.traverseMis(missionGroup, element); // Recurse
 			} else if (element._type === MissionElementType.InteriorInstance) {
-				this.interiorDependencies.add(element.interiorfile);
+				missionGroup.interiorDependencies.add(element.interiorfile);
 			} else if (element._type === MissionElementType.PathedInterior) {
-				this.interiorDependencies.add(element.interiorresource);
+				missionGroup.interiorDependencies.add(element.interiorresource);
 			} else if (element._type === MissionElementType.Sky) {
-				this.skyDependencies.add(element.materiallist);
+				missionGroup.skyDependencies.add(element.materiallist);
 			} else if (element._type === MissionElementType.TSStatic) {
-				this.shapeDependencies.add(element.shapename);
+				missionGroup.shapeDependencies.add(element.shapename);
 			} else if (element._type === MissionElementType.Trigger) {
 				// Check for music dependencies caused by MusicTriggers
 				if (element.datablock?.toLowerCase() === 'musictrigger') {
-					await this.registerDependency(path.posix.join('sound/music', element.text), 'extension-agnostic', this.misFilePath);
+					await this.registerDependency(missionGroup, path.posix.join('sound/music', element.text), 'extension-agnostic', missionGroup.misFilePath);
 				}
 			}
 		}
@@ -217,7 +241,7 @@ export class MissionUpload {
 	 * be found in the regular PQ data directory and if so, we don't need to worry about it. If, however, it can't even be found there, then
 	 * this mission is dependening on an asset that cannot be found and therefore, the mission can't be submitted.
 	 */
-	async registerDependency(dependency: string, matchType: 'exact' | 'extension-agnostic', requiredBy: string, permittedExtensions?: string[], interiorsMbgFallback = false, originalDependency?: string): Promise<{ found: jszip.JSZipObject, relativeDirectory: string }> {;
+	async registerDependency(group: MissionGroup, dependency: string, matchType: 'exact' | 'extension-agnostic', requiredBy: string, permittedExtensions?: string[], interiorsMbgFallback = false, originalDependency?: string): Promise<{ found: jszip.JSZipObject, relativeDirectory: string }> {;
 		// First, disect the dependency part into its different piecess
 		let fileName = Util.getFileName(dependency);
 		let dataIndex = dependency.indexOf('data/');
@@ -243,7 +267,7 @@ export class MissionUpload {
 				if (!found) {
 					// The dependency was found in the archive, so add it to the normalized directory
 					found = this.zip.files[name];
-					this.normalizedDirectory.set(path.posix.join(relativeDirectory, fileName2), this.zip.files[name]);
+					group.normalizedDirectory.set(path.posix.join(relativeDirectory, fileName2), this.zip.files[name]);
 				} else {
 					// The dependency was found more than once and we aren't sure which one is meant, so abort.
 					this.problems.add(`The dependency ${dependency} could not be uniquely resolved to a file in the archive.`);
@@ -263,7 +287,7 @@ export class MissionUpload {
 			if (!fullPath) { // We couldn't resolve the dependency to any file
 				if (interiorsMbgFallback && relativeDirectory.startsWith('interiors/')) {
 					// Special case: If the directory is interiors, check again in interiors_mbg. This is because the default MBG assets (in PQ) are located in interiors_mbg, but the paths within the .mis files still point to interiors (it does some directory shadowing).
-					return this.registerDependency(dependency.replace('interiors/', 'interiors_mbg/'), matchType, requiredBy, permittedExtensions, false, dependency);
+					return this.registerDependency(group, dependency.replace('interiors/', 'interiors_mbg/'), matchType, requiredBy, permittedExtensions, false, dependency);
 				} else {
 					// We definitely couldn't locate a file, add a problem
 					this.problems.add(`Missing dependency: ${originalDependency ?? dependency} is required by ${requiredBy} but couldn't be found.`);
@@ -279,9 +303,9 @@ export class MissionUpload {
 	}
 
 	/** Walks over all interiors and checks their dependencies. */
-	async checkInteriors() {
-		for (let dependency of this.interiorDependencies) {
-			let result = await this.registerDependency(dependency, 'exact', this.misFilePath, null, true);
+	async checkInteriors(group: MissionGroup) {
+		for (let dependency of group.interiorDependencies) {
+			let result = await this.registerDependency(group, dependency, 'exact', group.misFilePath, null, true);
 			if (!result) continue;
 
 			let { found, relativeDirectory } = result;
@@ -292,20 +316,23 @@ export class MissionUpload {
 
 			// Go over all materials and add them as dependencies
 			for (let interior of dif.interiors.concat(dif.subObjects)) {
-				for (let material of interior.materialList) {
+				let usedMaterials = new Set(interior.surfaces.map(x => x.textureIndex));
+
+				for (let [i, material] of interior.materialList.entries()) {
 					if (IGNORE_MATERIALS.includes(material)) continue;
+					if (!usedMaterials.has(i)) continue; // If the material isn't used in any surface, we don't care if it's there or not
 
 					let materialFileName = Util.getFileName(material);
-					await this.registerDependency(path.posix.join(relativeDirectory, materialFileName), 'extension-agnostic', dependency, IMAGE_EXTENSIONS);
+					await this.registerDependency(group, path.posix.join(relativeDirectory, materialFileName), 'extension-agnostic', dependency, IMAGE_EXTENSIONS);
 				}
 			}
 		}
 	}
 
 	/** Walks over the sky dependencies. */
-	async checkSky() {
-		for (let dependency of this.skyDependencies) {
-			let result = await this.registerDependency(dependency, 'exact', this.misFilePath);
+	async checkSky(group: MissionGroup) {
+		for (let dependency of group.skyDependencies) {
+			let result = await this.registerDependency(group, dependency, 'exact', group.misFilePath);
 			if (!result) continue;
 
 			let { found, relativeDirectory } = result;
@@ -316,15 +343,15 @@ export class MissionUpload {
 	
 			// Register all sky textures as dependencies
 			for (let line of lines) {
-				await this.registerDependency(path.posix.join(relativeDirectory, line), 'extension-agnostic', dependency, IMAGE_EXTENSIONS);
+				await this.registerDependency(group, path.posix.join(relativeDirectory, line), 'extension-agnostic', dependency, IMAGE_EXTENSIONS);
 			}
 		}
 	}
 
 	/** Walks over all shapes and checks their dependencies. */
-	async checkShapes() {
-		for (let dependency of this.shapeDependencies) {
-			let result = await this.registerDependency(dependency, 'exact', this.misFilePath);
+	async checkShapes(group: MissionGroup) {
+		for (let dependency of group.shapeDependencies) {
+			let result = await this.registerDependency(group, dependency, 'exact', group.misFilePath);
 			if (!result) continue;
 
 			let { found, relativeDirectory } = result;
@@ -335,7 +362,7 @@ export class MissionUpload {
 
 			// Go over all materials and register them as dependencies
 			for (let matName of dtsFile.matNames) {
-				let result2 = await this.registerDependency(path.posix.join(relativeDirectory, matName), 'extension-agnostic', dependency, IMAGE_EXTENSIONS.concat(['.ifl']));
+				let result2 = await this.registerDependency(group, path.posix.join(relativeDirectory, matName), 'extension-agnostic', dependency, IMAGE_EXTENSIONS.concat(['.ifl']));
 				if (!result2) continue;
 
 				let found2 = result2.found;
@@ -347,7 +374,7 @@ export class MissionUpload {
 					for (let line of lines) {
 						let textureName = line.split(' ')[0]?.trim();
 						if (textureName) {
-							await this.registerDependency(path.posix.join(relativeDirectory, textureName), 'extension-agnostic', found2.name, IMAGE_EXTENSIONS);
+							await this.registerDependency(group, path.posix.join(relativeDirectory, textureName), 'extension-agnostic', found2.name, IMAGE_EXTENSIONS);
 						}
 					}
 				}
@@ -359,37 +386,86 @@ export class MissionUpload {
 		return Date.now() - this.started >= UPLOAD_TTL;
 	}
 
-	/** Submits the uploaded mission to the database and writes its normalized directory to disk. */
-	async submit(submitter: number, remarks?: string) {
+	/** Submits the uploaded missions to the database and writes their normalized directories to disk. */
+	async submit(submitter: AccountDoc, requestBody: {
+		remarks: string[],
+		addToPacks: number[],
+		newPack?: {
+			name: string,
+			description: string
+		}
+	}) {
 		if (this.hasExpired()) throw new Error("Expired.");
 
-		let levelId = keyValue.get('levelId');
-		keyValue.set('levelId', levelId + 1);
+		let packDocs: PackDoc[] = [];
+		let newPackId: number = null;
 
-		let directoryPath = path.join(__dirname, 'storage', 'levels', levelId.toString());
-		await fs.ensureDir(directoryPath); // Create the new directory
+		// Get all the pack docs ready
+		for (let packId of new Set(requestBody.addToPacks)) {
+			let doc = await db.packs.findOne({ _id: packId }) as PackDoc;
+			if (!doc) {
+				throw new Error("Pack doesn't exist.");
+			}
 
-		// Write everything to disk
-		for (let [relativePath, file] of this.normalizedDirectory) {
-			let buffer = await file.async('nodebuffer');
-			let fullPath = path.join(directoryPath, relativePath);
-			await fs.ensureFile(fullPath);
-			await fs.writeFile(fullPath, buffer);
+			packDocs.push(doc);
 		}
 
-		// Create a mission at the newly created directory and hydrate it like we would with any other mission
-		let relativePath = [...this.normalizedDirectory.keys()].find(x => x.endsWith('.mis'));
-		let mission = new Mission(directoryPath, relativePath, levelId);
-		await mission.hydrate();
+		// Create a new pack if needed
+		if (requestBody.newPack) {
+			let newPackDoc = await createPack(submitter, requestBody.newPack.name, requestBody.newPack.description, false);
+			packDocs.push(newPackDoc);
+			newPackId = newPackDoc._id;
+		}
 
-		// Add the mission to the database
-		let doc = mission.createDoc();
-		doc.addedBy = submitter;
-		doc.remarks = remarks ?? '';
+		let docs: MissionDoc[] = [];
+		for (let [i, group] of this.groups.entries()) {
+			let levelId = keyValue.get('levelId');
+			keyValue.set('levelId', levelId + 1);
 
-		await db.missions.insert(doc);
+			let directoryPath = path.join(__dirname, 'storage', 'levels', levelId.toString());
+			await fs.ensureDir(directoryPath); // Create the new directory
+
+			// Write everything to disk
+			for (let [relativePath, file] of group.normalizedDirectory) {
+				let buffer = await file.async('nodebuffer');
+				let fullPath = path.join(directoryPath, relativePath);
+				await fs.ensureFile(fullPath);
+				await fs.writeFile(fullPath, buffer);
+			}
+
+			// Create a mission at the newly created directory and hydrate it like we would with any other mission
+			let relativePath = [...group.normalizedDirectory.keys()].find(x => x.endsWith('.mis'));
+			let mission = new Mission(directoryPath, relativePath, levelId);
+			await mission.hydrate();
+
+			// Add the mission to the database
+			let doc = mission.createDoc();
+			doc.addedBy = submitter._id;
+			doc.remarks = requestBody.remarks[i] ?? '';
+
+			docs.push(doc);
+		}
+
+		await db.missions.insert(docs);
+
+		// Now, let's do all the pack updating:
+
+		let promises: Promise<any>[] = [];
+
+		for (let packDoc of packDocs) {
+			packDoc.levels.push(...docs.map(x => x._id));
+			promises.push(createPackThumbnail(packDoc));
+
+			await db.packs.update({ _id: packDoc._id }, packDoc, { upsert: true });
+		}
+
+		await Promise.allSettled(promises); // Wait for all the thumbnails to finish
+
 		ongoingUploads.delete(this.id);
 
-		return doc;
+		return {
+			docs,
+			newPackId
+		};
 	}
 }